Initial Commit, including OAuth
This is the initial commit of this repo, this app, this modest slice of
document storage, processing, and preservation. Long live Docput, be it under
whatever name we come to know it.

This is Phoenix template with a lot wholly and shamelessly stolen from the much
better constable.io (https://github.com/thoughtbot/constable). Specially, the
"all of the useful bits" is cribbed.