YAML, Mustache, etc. is all working
THIS COMMIT IS WAY TOO BIG

Problem is that I got ahead of myself in tinkering and trying to figure stuff
out and now I can't really separate it all into nice commits. But, basically,
this does it all. You have layouts that have documents/revisions. When you
make a revision it does all the rendering and stashes that in the DB. It does
all the YAML frontmatter parsing. It does all the Mustache parsing. It does
all the Markdown parsing.